kallsyms: Add vmlinux kallsyms parsing
If kallsyms symbols are found in the vmcoreinfo, and DWARF info is not
found, then we can try to parse the kallsyms to provide symbol data.
This commit adds a kernel_info system which will contain any internal
debuginfo, and adds a kallsyms registry as the first component of the
kernel info. At initialization time, when DWARF is not available, parse
the kallsyms into an array of symbols, addresses, and symbol types.

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>