Refactor ELF symbol search into Symbol Finder API
Symbol lookup is not modular, like type or object lookup. To enable
kallsyms as a source for systems, we need a pluggable Symbol Finder API.
Define this symbol finder API as a flexible search which allows looking
up one or more symbols, listing all symbols, as well as conditioning
based on name, address, or both.

Refactor the ELF symbol search to fit this API, and make the existing
symbol lookups based on top of this. Leave
drgn_program_find_symbol_by_address_internal() alone -- at least for
now. It includes a Dwfl_Module argument which can't be provided to the
Symbol Finder API.

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>