cli: Set path when importing a script
When running a python script with the drgn cli, an import of
a module in the current directory does not find that module.
This is surprising for regular python users.

According to the python documentation, sys.path is modified when
a script is passed on the command line:

  "If the script name refers directly to a Python file, the
   directory containing that file is added to the start of
   sys.path, and the file is executed as the __main__ module."

However, it does not set the path if the passed in script is a
zipfile. Use pkgutil.get_importer() to check if this is the case
and only add the path if it returns None.

Add the same operation in drgn so that importing modules from
the same directory as the script work as expected.

Link: https://docs.python.org/3/using/cmdline.html#using-on-interface-options
Signed-off-by: Logan Gunthorpe <logang@deltatee.com>