tests: fix race in tests using fork_and_pause()+proc_blocked()
Every once in awhile, a tests.linux_kernel.test_stack_trace test fails
with a "cannot unwind stack of running task" error or without being able
to find pause or ppoll in the trace. I previously attempted to fix this
in commit a5845e63d48d ("tests: fix race condition in stack trace
tests"). However, now I'm seeing the forked process try to open
/etc/ld.so.cache before pausing for some reason. Since this can block,
proc_blocked() might return true before the process is actually in
pause(). If we get a stack trace while the process is in the wrong
syscall or in between calling another syscall and pause(), we will fail
as mentioned above.

Fix it in a few parts:

1. Use sigwait() instead of pause(), which I doubt is used anywhere else
   while forking and won't get woken up by stray signals.
2. Wait until /proc/$pid/syscall shows that we're blocked in
   rt_sigtimedwait or rt_sigtimedwait_time64 specifically.
3. Replace the manual fork_and_pause(), wait_until(proc_blocked, pid),
   os.kill(pid, signal.SIGKILL), os.waitpid(pid, 0) sequence with a
   context manager that takes care of all of that, fork_and_sigwait().

Signed-off-by: Omar Sandoval <osandov@osandov.com>