cli: print nicer messages for common errors
Michel pointed out that running drgn with no arguments without sudo
prints a verbose error:

  $ drgn
  drgn 0.0.21 (using Python 3.10.8, elfutils 0.187, with libkdumpfile)
  Traceback (most recent call last):
    File "/usr/bin/drgn", line 33, in <module>
      sys.exit(load_entry_point('drgn==0.0.21', 'console_scripts', 'drgn')())
    File "/usr/lib/python3.10/site-packages/drgn/internal/cli.py", line 168, in main
      prog.set_kernel()
  PermissionError: [Errno 13] Permission denied: '/proc/kcore'

In fact, any errors that we get when initializing the drgn.Program
result in an ugly traceback. Let's print specific error messages for the
most common errors:

- Trying to debug the live kernel without root.
- Trying to debug a live process without ptrace attach permissions.

There are a lot of details for both of these that could use elaboration,
but that should probably go in the documentation.

Also format other common errors without the traceback.

Suggested-by: Michel Alexandre Salim <michel@michel-slm.name>
Signed-off-by: Omar Sandoval <osandov@osandov.com>