helpers: add helper to iterate over all slabs of a slab-cache.
Having dedicated helpers to iterate over all slabs of a slab-cache,
not only helps in current use case(s) (i.e.iterating over objects
of a slab-cache) but it also helps in cases where we intend to
perform sanity checks on individual slabs.
For example we can iterate over all slabs to check that number of
objects are within allowed limit or to check that number of inuse
objects is not more than total number of objects on the slab.

Signed-off-by: Imran Khan <imran.f.khan@oracle.com>