helpers: slab: Rename _page_objects to _page_allocated_objects.
Right now _page_objects returns allocated objects of the slab for
both SLUB and SLAB. In certain debug scenarios we may be interested
in all or only free objects of a slab.
Next changes are introducing such  helpers.
So rename _page_objects to _page_allocated_objects, to indicate its
current functionality and also to allow addition of other helpers.

Signed-off-by: Imran Khan <imran.f.khan@oracle.com>