cli: make missing debug info warning more prominent
I've had lots of reports that turned out to be due to missing debug info
(e.g., #161, #175, #231, and more offline). A big part of the problem is
probably that the warning gets lost among the startup information. Let's
move the warning to be the last thing we print and also color it.

After the debug info discovery rework, we can print more specific
advice, but this is better than nothing.

Closes #41.

Signed-off-by: Omar Sandoval <osandov@osandov.com>