vmtest: kbuild: enable net_cls cgroups
For the sock_cgroup_ptr() test case in #255, we need
CONFIG_SOCK_CGROUP_DATA to be selected. Since Linux 4.10, we get that
from CONFIG_CGROUP_BPF, but before that, we need either
CONFIG_CGROUP_NET_CLASSID or CONFIG_CGROUP_NET_PRIO. The former appears
to be slightly smaller than the latter, so let's enable that one.

Signed-off-by: Omar Sandoval <osandov@osandov.com>