Index classes as type namespaces
In C++, a class definition creates a namespace. These are now
indexed in the same fashion as normal `namespace`s, enabling nested
types within them to be found as expected with `drgn.Program.type()`.

For the time being, this is done solely for classes and not for structs,
as the namespacing behavior of nesting types within structs differs
between C and C++, whereas classes exist only in C++.

Signed-off-by: Kevin Svetlitski <svetlitski@meta.com>