contrib: ptdrgn: Improve Object completion
Completion of drgn Objects *is* already possible in ptpython, but
unfortunately the list of completions is alphabetical (except for
underscored attributes, which are last). This means that default Object
fields like "absent_" and "type_" are not distinguished from the Object
member fields.

Resolve this by detecting when the first completion is "absent_" and
then moving all the non-default completions to the beginning of the
list. The result is that Object members are enumerated first, followed
by the default Object fields.

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>