libdrgn: program: find crashed task with cpu_curr() instead of find_task()
s390x populates the pid field in NT_PRSTATUS with the CPU number plus 1
[1] instead of the PID of the task that was running on that CPU. This
means that we get the wrong task_struct from drgn_program_find_thread()
in drgn_program_kernel_core_dump_cache_crashed_thread(), or don't find
the task_struct and crash because of a missing NULL check.

We can work around this and also gracefully handle the normal and idle
cases by instead getting the current task_struct from the CPU runqueue.
This is slightly racy: rq->curr is updated in __schedule() [2] before
the registers and stack are switched in context_switch() [3]. However,
it was already racy, since the pid field in NT_PRSTATUS is populated
from current, which is updated after the registers and stack are
switched (at least on x86-64) [4].

Closes #314.

1: https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/tree/arch/s390/kernel/crash_dump.c?h=v6.4#n309
2: https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/tree/kernel/sched/core.c?h=v6.4#n6646
3: https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/tree/kernel/sched/core.c?h=v6.4#n5343
4: https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/tree/arch/x86/kernel/process_64.c?h=v6.4#n621

Signed-off-by: Omar Sandoval <osandov@osandov.com>