helpers: cpumask: add helper to convert a cpumask to string of cpu ranges.
In some cases (for example getting affinity of an irq), it is better
to have an easily understandable list of cpus corresonding to a given
cpumask.
This helper converts a given cpumask to string, such that the string
represents the range of CPUs that are present in the given mask.

Signed-off-by: Imran Khan <imran.f.khan@oracle.com>